+++ crypto/hashing.go
@@ -1,4 +1,4 @@
-package main
+package crypto
 
 import (
 	"encoding/hex"
